move led and lights into hardware_components
