--- susi_linux/hardware_components/__init__.py
@@ -1,1 +1,0 @@
-from .rpi_wake_button import RaspberryPiWakeButton
