+++ susi_linux/hardware_components/lights.py
