--- susi_linux/light_control/white_light_control.py
@@ -1,7 +1,0 @@
-""" This script has been created to control the random lights that are started
-on the ReSpeaker and also to add custom light animations in the future
-"""
-
-from ..lights import lights
-
-lights.off()
